multiagent booking system/requirements.txt
langgraph
langchain-openai
python-dotenv
fastapi
uvicorn
pydantic
httpx
asyncio
typing-extensions
# MCP — Model Context Protocol (for flight database tool)
mcp
fastmcp
